# openapi.yaml
openapi: 3.0.1
info:
    description: |
        VICOO 个人知识管理助手 API
        
        提供笔记管理、知识图谱、AI问答、标签系统等功能，
        服务于 AI 智能助手的全局知识库问答能力。
    title: VICOO 知识管理助手插件
    version: v1
servers:
    - url: http://localhost:8000
      description: 本地开发服务器
paths:
...
    /api/tags:
        get:
            operationId: list_tags
            summary: 获取标签列表
            responses:
                "200":
                    description: 成功
                    content:
                        application/json:
                            schema:
                                type: object
        post:
            operationId: create_tag
            summary: 创建标签
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                color:
                                    type: string
            responses:
                "200":
                    description: 成功
                    content:
                        application/json:
                            schema:
                                type: object
    /api/tags/{id}:
        delete:
            operationId: delete_tag
            summary: 删除标签
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: 成功
                    content:
                        application/json:
                            schema:
                                type: object

    # ==================== 知识图谱 ====================
...
    /api/analytics/activity:
        get:
            operationId: analytics_activity
            summary: 活动数据
            parameters:
                - name: days
                  in: query
                  schema:
                      type: integer
            responses:
                "200":
                    description: 成功
                    content:
                        application/json:
                            schema:
                                type: object

    # ==================== 设置 ====================
...